refactor: move app files to Main folder, move footer to src/ folder
