+++ package.json
@@ -6,8 +6,9 @@
     "flexboxgrid2": "^7.1.0",
     "normalize.css": "^8.0.0",
     "react": "^16.2.0",
-    "styled-components": "^3.1.6",
-    "react-dom": "^16.2.0"
+    "react-dom": "^16.2.0",
+    "react-router-dom": "^4.2.2",
+    "styled-components": "^3.1.6"
   },
   "scripts": {
     "start": "react-scripts start",
