+++ src/Main/Destinations/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import media from "../common/media";
+import media from "../../common/media";
 import Heading from "./Heading";
 import CityCard from "./CityCard";
 import Category from "./Category";
