+++ src/Main/Header/SearchForm/Dates/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import media from "../../../common/media";
+import media from "../../../../common/media";
 import DepartureDate from "./DepartureDate";
 import ArrivalDate from "./ArrivalDate";
