+++ src/Main/Header/SearchForm/PassengerAndClass.js
@@ -1,5 +1,5 @@
 import React from "react";
-import media from "../../common/media";
+import media from "../../../common/media";
 import styled from "styled-components";
 import arrowDown from "./icons/arrow_down.svg";
