+++ src/Main/Slider/index.js
@@ -1,7 +1,7 @@
 import React from "react";
-import media from "../common/media";
+import media from "../../common/media";
 import styled from "styled-components";
-import Button from "../common/Button";
+import Button from "../../common/Button";
 import aeroflot from "./img/aeroflot.png";
 import s7airlines from "./img/s7airlines.png";
 import one2trip from "./img/one2trip.png";
